+++ core/decision_func.go
