+++ core/estimation_func.go
