+++ core/ordering_func.go
