+++ core/target_func.go
