+++ core/wait_func.go
